Robustness Checks - Read description below!
Using CascaldiGarcia for unconditional responses and variables in logdifference:
- Results are robust using 1,2,3, or 4 principal components
- Results are robust using 1,2,3, or 4 lags of control variables
- Using SPF industrial production noise shocks are not anymore deflationary. However, all the other responses look similar.
- Using SPF real investment resuls are not good. Series is too short.
- Using different lead or lags of TFP does not particularly change the result. (lead > 8 and lag > 1 are the minimum requirements)

Results are robust for variables in loglevels, no cumsum

Results are robust for variable in logOneSidedHPfilter, no cumsum

Results are similar but not deflationary in logHPfilter, no sumsum

Results are not good when using logdiff of either SP500 or MI1Y. Probably I am doing something wrong but so far we do not get anything interesting.

Still need to apply the Livingston Survey and BandPass filter...We need to talk about that.